fully functional original algorithm

+++ BalloonTreeLayoutPlugin/src/net/phreakocious/balloontreelayout/BalloonTreeNodeLayoutData.java
@@ -8,8 +8,8 @@ import org.gephi.graph.spi.LayoutData;
  * @author phreakocious
  */
 public class BalloonTreeNodeLayoutData implements LayoutData {
-        public int d = 0;           // Constant distance used to position each child node's layout circle
-        public int r = 0;           // Radius of the layout circle containing the full subtree induced by this node
+        public double d = 0;         // Constant distance used to position each child node's layout circle
+        public double r = 0;         // Radius of the layout circle containing the full subtree induced by this node
         public double rx = 0;       // X coordinate, relative to barycenter
         public double ry = 0;       // Y coordinate, relative to barycenter
         public double a = 0;        // Angle equal to the halfsector formed by this node, relative to its ancestor
